+++ CHANGELOG.md
@@ -36,7 +36,8 @@ Expected: June 2021
 
 * Yang Deviation/deviate [deviation statement not yet support #211](https://github.com/clicon/clixon/issues/211)
   * See RFC7950 Sec 5.6.3
-  * Work-in-progress
+  * Implemented: "not-supported" and "add"
+  * Not yet: "replace" and "delete"
 
 ### Minor features